third commit

+++ ipa-1.py
@@ -27,10 +27,10 @@ def savings(gross_pay, tax_rate, expenses):
     # Replace `pass` with your code.
     # Stay within the function. Only use the parameters as input. The function should return your answer.
 
-    tax = (gross_pay * tax_rate)
+    tax = int(gross_pay * tax_rate)
     net_pay=gross_pay-tax
     after_tax_pay = net_pay - expenses
-    return int(after_tax_pay)
+    return (after_tax_pay)
 
 money_remaining=savings(12000,0.2,32)
 print("Remaining money:", money_remaining,"centavos")
